hackfest-25.2/image.bak/deps/harmony-service-lib-py/requirements.txt
boto3 ~= 1.14
deprecation ~= 2.1.0
pynacl ~= 1.4
pystac >= 1.0.0
python-json-logger ~= 2.0.1
requests ~= 2.24
urllib3 ~= 1.26.9
